REFACTOR :  store.js 이름 변경
