+++ src/main.js
@@ -1,7 +1,7 @@
 import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router/router.js";
-import store from './store/index.js';
+import store from './store/store.js';
 
 // import 'bootstrap'
 // import 'bootstrap/dist/css/bootstrap.min.css'
